Refactoring directory.
